Struct aws_sdk_costexplorer::client::fluent_builders::GetTags

pub struct GetTags<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> GetTags<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<GetTagsOutput, SdkError<GetTagsError>> where
R::Policy: SmithyRetryPolicy<GetTagsInputOperationOutputAlias, GetTagsOutput, GetTagsError, GetTagsInputOperationRetryAlias>,

pub fn search_string(self, inp: impl Into<String>) -> Self

The value that you want to search for.

pub fn set_search_string(self, input: Option<String>) -> Self

pub fn time_period(self, inp: DateInterval) -> Self

The start and end dates for retrieving the dimension values. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01.

pub fn set_time_period(self, input: Option<DateInterval>) -> Self

pub fn tag_key(self, inp: impl Into<String>) -> Self

The key of the tag that you want to return values for.

pub fn set_tag_key(self, input: Option<String>) -> Self

pub fn filter(self, inp: Expression) -> Self

Use Expression to filter by cost or by usage. There are two patterns:

Simple dimension values - You can set the dimension name and values for the filters that you plan to use. For example, you can filter for REGION==us-east-1 OR REGION==us-west-1. For GetRightsizingRecommendation, the Region is a full name (for example, REGION==US East (N. Virginia). The Expression example is as follows:

{ "Dimensions": { "Key": "REGION", "Values": [ "us-east-1", “us-west-1” ] } }

The list of dimension values are OR'd together to retrieve cost or usage data. You can create Expression and DimensionValues objects using either with* methods or set* methods in multiple lines.

Compound dimension values with logical operations - You can use multiple Expression types and the logical operators AND/OR/NOT to create a list of one or more Expression objects. By doing this, you can filter on more advanced options. For example, you can filter on ((REGION == us-east-1 OR REGION == us-west-1) OR (TAG.Type == Type1)) AND (USAGE_TYPE != DataTransfer). The Expression for that is as follows:

{ "And": [ {"Or": [ {"Dimensions": { "Key": "REGION", "Values": [ "us-east-1", "us-west-1" ] }}, {"Tags": { "Key": "TagName", "Values": ["Value1"] } } ]}, {"Not": {"Dimensions": { "Key": "USAGE_TYPE", "Values": ["DataTransfer"] }}} ] }

Because each Expression can have only one operator, the service returns an error if more than one is specified. The following example shows an Expression object that creates an error.

{ "And": [ ... ], "DimensionValues": { "Dimension": "USAGE_TYPE", "Values": [ "DataTransfer" ] } }

For the GetRightsizingRecommendation action, a combination of OR and NOT isn't supported. OR isn't supported between different dimensions, or dimensions and tags. NOT operators aren't supported. Dimensions are also limited to LINKED_ACCOUNT, REGION, or RIGHTSIZING_TYPE.

For the GetReservationPurchaseRecommendation action, only NOT is supported. AND and OR aren't supported. Dimensions are limited to LINKED_ACCOUNT.

pub fn set_filter(self, input: Option<Expression>) -> Self

pub fn sort_by(self, inp: impl Into<SortDefinition>) -> Self

Appends an item to SortBy.

To override the contents of this collection use set_sort_by.

The value by which you want to sort the data.

The key represents cost and usage metrics. The following values are supported:

BlendedCost

UnblendedCost

AmortizedCost

NetAmortizedCost

NetUnblendedCost

UsageQuantity

NormalizedUsageAmount

Supported values for SortOrder are ASCENDING or DESCENDING.

When using SortBy, NextPageToken and SearchString are not supported.

pub fn set_sort_by(self, input: Option<Vec<SortDefinition>>) -> Self

pub fn max_results(self, inp: i32) -> Self

This field is only used when SortBy is provided in the request. The maximum number of objects that to be returned for this request. If MaxResults is not specified with SortBy, the request will return 1000 results as the default value for this parameter.

For GetTags, MaxResults has an upper limit of 1000.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn next_page_token(self, inp: impl Into<String>) -> Self

The token to retrieve the next set of results. Amazon Web Services provides the token when the response from a previous call has more results than the maximum page size.

pub fn set_next_page_token(self, input: Option<String>) -> Self