Struct aws_sdk_costexplorer::client::fluent_builders::GetUsageForecast

pub struct GetUsageForecast<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> GetUsageForecast<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<GetUsageForecastOutput, SdkError<GetUsageForecastError>> where
R::Policy: SmithyRetryPolicy<GetUsageForecastInputOperationOutputAlias, GetUsageForecastOutput, GetUsageForecastError, GetUsageForecastInputOperationRetryAlias>,

pub fn time_period(self, inp: DateInterval) -> Self

The start and end dates of the period that you want to retrieve usage forecast for. The start date is inclusive, but the end date is exclusive. For example, if start is 2017-01-01 and end is 2017-05-01, then the cost and usage data is retrieved from 2017-01-01 up to and including 2017-04-30 but not including 2017-05-01. The start date must be equal to or later than the current date to avoid a validation error.

pub fn set_time_period(self, input: Option<DateInterval>) -> Self

pub fn metric(self, inp: Metric) -> Self

Which metric Cost Explorer uses to create your forecast.

Valid values for a GetUsageForecast call are the following:

USAGE_QUANTITY

NORMALIZED_USAGE_AMOUNT

pub fn set_metric(self, input: Option<Metric>) -> Self

pub fn granularity(self, inp: Granularity) -> Self

How granular you want the forecast to be. You can get 3 months of DAILY forecasts or 12 months of MONTHLY forecasts.

The GetUsageForecast operation supports only DAILY and MONTHLY granularities.

pub fn set_granularity(self, input: Option<Granularity>) -> Self

pub fn filter(self, inp: Expression) -> Self

The filters that you want to use to filter your forecast. The GetUsageForecast API supports filtering by the following dimensions:

AZ

INSTANCE_TYPE

LINKED_ACCOUNT

LINKED_ACCOUNT_NAME

OPERATION

PURCHASE_TYPE

REGION

SERVICE

USAGE_TYPE

USAGE_TYPE_GROUP

RECORD_TYPE

OPERATING_SYSTEM

TENANCY

SCOPE

PLATFORM

SUBSCRIPTION_ID

LEGAL_ENTITY_NAME

DEPLOYMENT_OPTION

DATABASE_ENGINE

INSTANCE_TYPE_FAMILY

BILLING_ENTITY

RESERVATION_ID

SAVINGS_PLAN_ARN

pub fn set_filter(self, input: Option<Expression>) -> Self

pub fn prediction_interval_level(self, inp: i32) -> Self

Cost Explorer always returns the mean forecast as a single point. You can request a prediction interval around the mean by specifying a confidence level. The higher the confidence level, the more confident Cost Explorer is about the actual value falling in the prediction interval. Higher confidence levels result in wider prediction intervals.

pub fn set_prediction_interval_level(self, input: Option<i32>) -> Self