Module aws_sdk_costexplorer::error

Modules

bill_expiration_exception

See BillExpirationException

data_unavailable_exception

See DataUnavailableException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

request_changed_exception

See RequestChangedException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

unknown_monitor_exception

See UnknownMonitorException

unknown_subscription_exception

See UnknownSubscriptionException

unresolvable_usage_unit_exception

See UnresolvableUsageUnitException

Structs

BillExpirationException

The requested report expired. Update the date interval and try again.

CreateAnomalyMonitorError

CreateAnomalySubscriptionError

CreateCostCategoryDefinitionError

DataUnavailableException

The requested data is unavailable.

DeleteAnomalyMonitorError

DeleteAnomalySubscriptionError

DeleteCostCategoryDefinitionError

DescribeCostCategoryDefinitionError

GetAnomaliesError

GetAnomalyMonitorsError

GetAnomalySubscriptionsError

GetCostAndUsageError

GetCostAndUsageWithResourcesError

GetCostCategoriesError

GetCostForecastError

GetDimensionValuesError

GetReservationCoverageError

GetReservationPurchaseRecommendationError

GetReservationUtilizationError

GetRightsizingRecommendationError

GetSavingsPlansCoverageError

GetSavingsPlansPurchaseRecommendationError

GetSavingsPlansUtilizationDetailsError

GetSavingsPlansUtilizationError

GetTagsError

GetUsageForecastError

InvalidNextTokenException

The pagination token is invalid. Try again without a pagination token.

LimitExceededException

You made too many calls in a short period of time. Try again later.

ListCostCategoryDefinitionsError

ProvideAnomalyFeedbackError

RequestChangedException

Your request parameters changed between pages. Try again with the old parameters or

ResourceNotFoundException

ServiceQuotaExceededException

You've reached the limit on the number of resources you can create, or exceeded the

UnknownMonitorException

The cost anomaly monitor does not exist for the account.

UnknownSubscriptionException

The cost anomaly subscription does not exist for the account.

UnresolvableUsageUnitException

Cost Explorer was unable to identify the usage unit. Provide UsageType/UsageTypeGroup filter selections that contain matching units, for example: hours.

UpdateAnomalyMonitorError

UpdateAnomalySubscriptionError

UpdateCostCategoryDefinitionError

Enums

CreateAnomalyMonitorErrorKind

CreateAnomalySubscriptionErrorKind

CreateCostCategoryDefinitionErrorKind

DeleteAnomalyMonitorErrorKind

DeleteAnomalySubscriptionErrorKind

DeleteCostCategoryDefinitionErrorKind

DescribeCostCategoryDefinitionErrorKind

GetAnomaliesErrorKind

GetAnomalyMonitorsErrorKind

GetAnomalySubscriptionsErrorKind

GetCostAndUsageErrorKind

GetCostAndUsageWithResourcesErrorKind

GetCostCategoriesErrorKind

GetCostForecastErrorKind

GetDimensionValuesErrorKind

GetReservationCoverageErrorKind

GetReservationPurchaseRecommendationErrorKind

GetReservationUtilizationErrorKind

GetRightsizingRecommendationErrorKind

GetSavingsPlansCoverageErrorKind

GetSavingsPlansPurchaseRecommendationErrorKind

GetSavingsPlansUtilizationDetailsErrorKind

GetSavingsPlansUtilizationErrorKind

GetTagsErrorKind

GetUsageForecastErrorKind

ListCostCategoryDefinitionsErrorKind

ProvideAnomalyFeedbackErrorKind

UpdateAnomalyMonitorErrorKind

UpdateAnomalySubscriptionErrorKind

UpdateCostCategoryDefinitionErrorKind