Module aws_sdk_costexplorer::input

Modules

create_anomaly_monitor_input

See CreateAnomalyMonitorInput

create_anomaly_subscription_input

See CreateAnomalySubscriptionInput

create_cost_category_definition_input

See CreateCostCategoryDefinitionInput

delete_anomaly_monitor_input

See DeleteAnomalyMonitorInput

delete_anomaly_subscription_input

See DeleteAnomalySubscriptionInput

delete_cost_category_definition_input

See DeleteCostCategoryDefinitionInput

describe_cost_category_definition_input

See DescribeCostCategoryDefinitionInput

get_anomalies_input

See GetAnomaliesInput

get_anomaly_monitors_input

See GetAnomalyMonitorsInput

get_anomaly_subscriptions_input

See GetAnomalySubscriptionsInput

get_cost_and_usage_input

See GetCostAndUsageInput

get_cost_and_usage_with_resources_input

See GetCostAndUsageWithResourcesInput

get_cost_categories_input

See GetCostCategoriesInput

get_cost_forecast_input

See GetCostForecastInput

get_dimension_values_input

See GetDimensionValuesInput

get_reservation_coverage_input

See GetReservationCoverageInput

get_reservation_purchase_recommendation_input

See GetReservationPurchaseRecommendationInput

get_reservation_utilization_input

See GetReservationUtilizationInput

get_rightsizing_recommendation_input

See GetRightsizingRecommendationInput

get_savings_plans_coverage_input

See GetSavingsPlansCoverageInput

get_savings_plans_purchase_recommendation_input

See GetSavingsPlansPurchaseRecommendationInput

get_savings_plans_utilization_details_input

See GetSavingsPlansUtilizationDetailsInput

get_savings_plans_utilization_input

See GetSavingsPlansUtilizationInput

get_tags_input

See GetTagsInput

get_usage_forecast_input

See GetUsageForecastInput

list_cost_category_definitions_input

See ListCostCategoryDefinitionsInput

provide_anomaly_feedback_input

See ProvideAnomalyFeedbackInput

update_anomaly_monitor_input

See UpdateAnomalyMonitorInput

update_anomaly_subscription_input

See UpdateAnomalySubscriptionInput

update_cost_category_definition_input

See UpdateCostCategoryDefinitionInput

Structs

CreateAnomalyMonitorInput

CreateAnomalySubscriptionInput

CreateCostCategoryDefinitionInput

DeleteAnomalyMonitorInput

DeleteAnomalySubscriptionInput

DeleteCostCategoryDefinitionInput

DescribeCostCategoryDefinitionInput

GetAnomaliesInput

GetAnomalyMonitorsInput

GetAnomalySubscriptionsInput

GetCostAndUsageInput

GetCostAndUsageWithResourcesInput

GetCostCategoriesInput

GetCostForecastInput

GetDimensionValuesInput

GetReservationCoverageInput

You can use the following request parameters to query for how much of your instance usage a reservation covered.

GetReservationPurchaseRecommendationInput

GetReservationUtilizationInput

GetRightsizingRecommendationInput

GetSavingsPlansCoverageInput

GetSavingsPlansPurchaseRecommendationInput

GetSavingsPlansUtilizationDetailsInput

GetSavingsPlansUtilizationInput

GetTagsInput

GetUsageForecastInput

ListCostCategoryDefinitionsInput

ProvideAnomalyFeedbackInput

UpdateAnomalyMonitorInput

UpdateAnomalySubscriptionInput

UpdateCostCategoryDefinitionInput