Module aws_sdk_costexplorer::model

Modules

anomaly

See Anomaly

anomaly_date_interval

See AnomalyDateInterval

anomaly_monitor

See AnomalyMonitor

anomaly_score

See AnomalyScore

anomaly_subscription

See AnomalySubscription

cost_category

See CostCategory

cost_category_inherited_value_dimension

See CostCategoryInheritedValueDimension

cost_category_processing_status

See CostCategoryProcessingStatus

cost_category_reference

See CostCategoryReference

cost_category_rule

See CostCategoryRule

cost_category_split_charge_rule

See CostCategorySplitChargeRule

cost_category_split_charge_rule_parameter

See CostCategorySplitChargeRuleParameter

cost_category_values

See CostCategoryValues

coverage

See Coverage

coverage_by_time

See CoverageByTime

coverage_cost

See CoverageCost

coverage_hours

See CoverageHours

coverage_normalized_units

See CoverageNormalizedUnits

current_instance

See CurrentInstance

date_interval

See DateInterval

dimension_values

See DimensionValues

dimension_values_with_attributes

See DimensionValuesWithAttributes

disk_resource_utilization

See DiskResourceUtilization

ebs_resource_utilization

See EbsResourceUtilization

ec2_instance_details

See Ec2InstanceDetails

ec2_resource_details

See Ec2ResourceDetails

ec2_resource_utilization

See Ec2ResourceUtilization

ec2_specification

See Ec2Specification

elasti_cache_instance_details

See ElastiCacheInstanceDetails

es_instance_details

See EsInstanceDetails

expression

See Expression

forecast_result

See ForecastResult

group

See Group

group_definition

See GroupDefinition

impact

See Impact

instance_details

See InstanceDetails

metric_value

See MetricValue

modify_recommendation_detail

See ModifyRecommendationDetail

network_resource_utilization

See NetworkResourceUtilization

rds_instance_details

See RdsInstanceDetails

redshift_instance_details

See RedshiftInstanceDetails

reservation_aggregates

See ReservationAggregates

reservation_coverage_group

See ReservationCoverageGroup

reservation_purchase_recommendation

See ReservationPurchaseRecommendation

reservation_purchase_recommendation_detail

See ReservationPurchaseRecommendationDetail

reservation_purchase_recommendation_metadata

See ReservationPurchaseRecommendationMetadata

reservation_purchase_recommendation_summary

See ReservationPurchaseRecommendationSummary

reservation_utilization_group

See ReservationUtilizationGroup

resource_details

See ResourceDetails

resource_utilization

See ResourceUtilization

result_by_time

See ResultByTime

rightsizing_recommendation

See RightsizingRecommendation

rightsizing_recommendation_configuration

See RightsizingRecommendationConfiguration

rightsizing_recommendation_metadata

See RightsizingRecommendationMetadata

rightsizing_recommendation_summary

See RightsizingRecommendationSummary

root_cause

See RootCause

savings_plans_amortized_commitment

See SavingsPlansAmortizedCommitment

savings_plans_coverage

See SavingsPlansCoverage

savings_plans_coverage_data

See SavingsPlansCoverageData

savings_plans_details

See SavingsPlansDetails

savings_plans_purchase_recommendation

See SavingsPlansPurchaseRecommendation

savings_plans_purchase_recommendation_detail

See SavingsPlansPurchaseRecommendationDetail

savings_plans_purchase_recommendation_metadata

See SavingsPlansPurchaseRecommendationMetadata

savings_plans_purchase_recommendation_summary

See SavingsPlansPurchaseRecommendationSummary

savings_plans_savings

See SavingsPlansSavings

savings_plans_utilization

See SavingsPlansUtilization

savings_plans_utilization_aggregates

See SavingsPlansUtilizationAggregates

savings_plans_utilization_by_time

See SavingsPlansUtilizationByTime

savings_plans_utilization_detail

See SavingsPlansUtilizationDetail

service_specification

See ServiceSpecification

sort_definition

See SortDefinition

subscriber

See Subscriber

tag_values

See TagValues

target_instance

See TargetInstance

terminate_recommendation_detail

See TerminateRecommendationDetail

total_impact_filter

See TotalImpactFilter

utilization_by_time

See UtilizationByTime

Structs

Anomaly

An unusual cost pattern. This consists of the detailed metadata and the current status

AnomalyDateInterval

The time period for an anomaly.

AnomalyMonitor

This object continuously inspects your account's cost data for anomalies. It's based

AnomalyScore

Quantifies the anomaly. The higher score means that it's more anomalous.

AnomalySubscription

The association between a monitor, threshold, and list of subscribers used to deliver

CostCategory

The structure of Cost Categories. This includes detailed metadata and the set of rules

CostCategoryInheritedValueDimension

When creating or updating a cost category, you can define the

CostCategoryProcessingStatus

The list of processing statuses for Cost Management products for a specific cost

CostCategoryReference

A reference to a Cost Category containing only enough information to identify the Cost

CostCategoryRule

Rules are processed in order. If there are multiple rules that match the line item,

CostCategorySplitChargeRule

Use the split charge rule to split the cost of one Cost Category value across several

CostCategorySplitChargeRuleParameter

The parameters for a split charge method.

CostCategoryValues

The Cost Categories values used for filtering the costs.

Coverage

The amount of instance usage that a reservation covered.

CoverageByTime

Reservation coverage for a specified period, in hours.

CoverageCost

How much it costs to run an instance.

CoverageHours

How long a running instance either used a reservation or was On-Demand.

CoverageNormalizedUnits

The amount of instance usage, in normalized units. You can use normalized units to see

CurrentInstance

Context about the current instance.

DateInterval

The time period of the request.

DimensionValues

The metadata that you can use to filter and group your results. You can use

DimensionValuesWithAttributes

The metadata of a specific type that you can use to filter and group your results. You

DiskResourceUtilization

The field that contains a list of disk (local storage) metrics that are associated

EbsResourceUtilization

The EBS field that contains a list of EBS metrics that are associated with the current

Ec2InstanceDetails

Details about the Amazon EC2 instances that Amazon Web Services recommends that you

Ec2ResourceDetails

Details on the Amazon EC2 Resource.

Ec2ResourceUtilization

Utilization metrics of the instance.

Ec2Specification

The Amazon EC2 hardware specifications that you want Amazon Web Services to provide

ElastiCacheInstanceDetails

Details about the Amazon ElastiCache instances that Amazon Web Services recommends that

EsInstanceDetails

Details about the Amazon ES instances that Amazon Web Services recommends that you

Expression

Use Expression to filter by cost or by usage. There are two patterns:

ForecastResult

The forecast that's created for your query.

Group

One level of grouped data in the results.

GroupDefinition

Represents a group when you specify a group by criteria or in the response to a query

Impact

The dollar value of the anomaly.

InstanceDetails

Details about the instances that Amazon Web Services recommends that you

MetricValue

The aggregated value for a metric.

ModifyRecommendationDetail

Details on the modification recommendation.

NetworkResourceUtilization

The network field that contains a list of network metrics that are associated with

RdsInstanceDetails

Details about the Amazon RDS instances that Amazon Web Services recommends that you

RedshiftInstanceDetails

Details about the Amazon Redshift instances that Amazon Web Services recommends that

ReservationAggregates

The aggregated numbers for your reservation usage.

ReservationCoverageGroup

A group of reservations that share a set of attributes.

ReservationPurchaseRecommendation

A specific reservation that Amazon Web Services recommends for purchase.

ReservationPurchaseRecommendationDetail

Details about your recommended reservation purchase.

ReservationPurchaseRecommendationMetadata

Information about this specific recommendation, such as the timestamp for when Amazon Web Services made a specific recommendation.

ReservationPurchaseRecommendationSummary

A summary about this recommendation, such as the currency code, the amount that

ReservationUtilizationGroup

A group of reservations that share a set of attributes.

ResourceDetails

Details on the resource.

ResourceUtilization

Resource utilization of current resource.

ResultByTime

The result that's associated with a time period.

RightsizingRecommendation

Recommendations to rightsize resources.

RightsizingRecommendationConfiguration

You can use RightsizingRecommendationConfiguration to customize

RightsizingRecommendationMetadata

Metadata for this recommendation set.

RightsizingRecommendationSummary

The summary of rightsizing recommendations

RootCause

The combination of Amazon Web Services service, linked account, Region, and usage type

SavingsPlansAmortizedCommitment

The amortized amount of Savings Plans purchased in a specific account during a

SavingsPlansCoverage

The amount of Savings Plans eligible usage that is covered by Savings Plans. All

SavingsPlansCoverageData

Specific coverage percentage, On-Demand costs, and spend covered by Savings Plans, and

SavingsPlansDetails

The attribute details on a specific Savings Plan.

SavingsPlansPurchaseRecommendation

Contains your request parameters, Savings Plan Recommendations Summary, and

SavingsPlansPurchaseRecommendationDetail

Details for your recommended Savings Plans.

SavingsPlansPurchaseRecommendationMetadata

Metadata about your Savings Plans Purchase Recommendations.

SavingsPlansPurchaseRecommendationSummary

Summary metrics for your Savings Plans Purchase Recommendations.

SavingsPlansSavings

The amount of savings that you're accumulating, against the public On-Demand rate of

SavingsPlansUtilization

The measurement of how well you're using your existing Savings Plans.

SavingsPlansUtilizationAggregates

The aggregated utilization metrics for your Savings Plans usage.

SavingsPlansUtilizationByTime

The amount of Savings Plans utilization, in hours.

SavingsPlansUtilizationDetail

A single daily or monthly Savings Plans utilization rate, and details for your

ServiceSpecification

Hardware specifications for the service that you want recommendations for.

SortDefinition

The details of how to sort the data.

Subscriber

The recipient of AnomalySubscription notifications.

TagValues

The values that are available for a tag.

TargetInstance

Details on recommended instance.

TerminateRecommendationDetail

Details on termination recommendation.

TotalImpactFilter

Filters cost anomalies based on the total impact.

UtilizationByTime

The amount of utilization, in hours.

Enums

AccountScope

AnomalyFeedbackType

AnomalySubscriptionFrequency

Context

CostCategoryInheritedValueDimensionName

CostCategoryRuleType

CostCategoryRuleVersion

The rule schema version in this particular Cost Category.

CostCategorySplitChargeMethod

CostCategorySplitChargeRuleParameterType

CostCategoryStatus

CostCategoryStatusComponent

Dimension

FindingReasonCode

Granularity

GroupDefinitionType

LookbackPeriodInDays

MatchOption

Metric

MonitorDimension

MonitorType

NumericOperator

OfferingClass

PaymentOption

PlatformDifference

RecommendationTarget

RightsizingType

SavingsPlansDataType

SortOrder

SubscriberStatus

SubscriberType

SupportedSavingsPlansType

TermInYears