Module aws_sdk_costexplorer::operation

Structs

CreateAnomalyMonitor

Creates a new cost anomaly detection monitor with the requested type and monitor

CreateAnomalySubscription

Adds a subscription to a cost anomaly detection monitor. You can use each subscription to

CreateCostCategoryDefinition

Creates a new Cost Category with the requested name and rules.

DeleteAnomalyMonitor

Deletes a cost anomaly monitor.

DeleteAnomalySubscription

Deletes a cost anomaly subscription.

DeleteCostCategoryDefinition

Deletes a Cost Category. Expenses from this month going forward will no longer be categorized with this Cost Category.

DescribeCostCategoryDefinition

Returns the name, ARN, rules, definition, and effective dates of a Cost Category that's defined in the account.

GetAnomalies

Retrieves all of the cost anomalies detected on your account during the time period that's

GetAnomalyMonitors

Retrieves the cost anomaly monitor definitions for your account. You can filter using a

GetAnomalySubscriptions

Retrieves the cost anomaly subscription objects for your account. You can filter using a

GetCostAndUsage

Retrieves cost and usage metrics for your account. You can specify which cost and

GetCostAndUsageWithResources

Retrieves cost and usage metrics with resources for your account. You can specify which cost and

GetCostCategories

Retrieves an array of Cost Category names and values incurred cost.

GetCostForecast

Retrieves a forecast for how much Amazon Web Services predicts that you will spend over the forecast time period that you select, based on your past costs.

GetDimensionValues

Retrieves all available filter values for a specified filter over a period of time. You can search the dimension values for an arbitrary string.

GetReservationCoverage

Retrieves the reservation coverage for your account. This enables you to see how much of your Amazon Elastic Compute Cloud, Amazon ElastiCache, Amazon Relational Database Service, or

GetReservationPurchaseRecommendation

Gets recommendations for which reservations to purchase. These recommendations could help you reduce your costs.

GetReservationUtilization

Retrieves the reservation utilization for your account. Management account in an organization have access to member accounts.

GetRightsizingRecommendation

Creates recommendations that

GetSavingsPlansCoverage

Retrieves the Savings Plans covered for your account. This enables you to see how much of your cost is covered by a Savings Plan. An organization’s management account can see the coverage of the associated member accounts. This supports dimensions, Cost Categories, and nested expressions. For any time period, you can filter data for Savings Plans usage with the following dimensions:

GetSavingsPlansPurchaseRecommendation

Retrieves your request parameters, Savings Plan Recommendations Summary and Details.

GetSavingsPlansUtilization

Retrieves the Savings Plans utilization for your account across date ranges with daily or monthly granularity. Management account in an organization have access to member accounts. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

GetSavingsPlansUtilizationDetails

Retrieves attribute data along with aggregate utilization and savings data for a given time period. This doesn't support granular or grouped data (daily/monthly) in response. You can't retrieve data by dates in a single response similar to GetSavingsPlanUtilization, but you have the option to make multiple calls to GetSavingsPlanUtilizationDetails by providing individual dates. You can use GetDimensionValues in SAVINGS_PLANS to determine the possible dimension values.

GetTags

Queries for available tag keys and tag values for a specified period. You can search the tag values for an arbitrary string.

GetUsageForecast

Retrieves a forecast for how much Amazon Web Services predicts that you will use over the forecast time period that you select, based on your past usage.

ListCostCategoryDefinitions

Returns the name, ARN, NumberOfRules and effective dates of all Cost Categories defined in the account. You have the option to use EffectiveOn to return a list of Cost Categories that were active on a specific date. If there is no EffectiveOn specified, you’ll see Cost Categories that are effective on the current date. If Cost Category is still effective, EffectiveEnd is omitted in the response. ListCostCategoryDefinitions supports pagination. The request can have a MaxResults range up to 100.

ProvideAnomalyFeedback

Modifies the feedback property of a given cost anomaly.

UpdateAnomalyMonitor

Updates an existing cost anomaly monitor. The changes made are applied going forward, and

UpdateAnomalySubscription

Updates an existing cost anomaly monitor subscription.

UpdateCostCategoryDefinition

Updates an existing Cost Category. Changes made to the Cost Category rules will be used to categorize the current month’s expenses and future expenses. This won’t change categorization for the previous months.